factoring ElboDeriv log prob calc functions to Model
