+++ src/ElboDeriv.jl
@@ -8,272 +8,17 @@ using ..Model
 import ..WCSUtils
 import ..SensitiveFloats.clear!
 
-
 # TODO make the functions dependent on these objects dependent on ..Model instead
-import ..Model: ElboArgs, ActivePixel #, SensitiveFloats
+import ..Model: ElboArgs, ActivePixel, ElboIntermediateVariables,
+                populate_fsm_vecs!,
+                BvnComponent, GalaxyCacheComponent, load_bvn_mixtures
 export ElboArgs, ActivePixel #, SensitiveFloats
+
+import ..SensitiveFloats: SensitiveFloat, zero_sensitive_float, add_sources_sf!
 
 include("elbo_kl.jl")
 include("source_brightness.jl")
 #include("bivariate_normals.jl")
-
-
-"""
-Convolve the current locations and galaxy shapes with the PSF.  If
-calculate_derivs is true, also calculate derivatives and hessians for
-active sources.
-
-Args:
- - psf: A vector of PSF components
- - ea: The current ElboArgs
- - b: The current band
- - calculate_derivs: Whether to calculate derivatives for active sources.
-
-Returns:
- - star_mcs: An array of BvnComponents with indices
-    - PSF component
-    - Source (index within active_sources)
- - gal_mcs: An array of BvnComponents with indices
-    - PSF component
-    - Galaxy component
-    - Galaxy type
-    - Source (index within active_sources)
-  Hessians are only populated for s in ea.active_sources.
-"""
-function load_bvn_mixtures{NumType <: Number}(
-    ea::ElboArgs{NumType}, b::Int;
-    calculate_derivs::Bool=true, calculate_hessian::Bool=true)
-
-  star_mcs = Array(BvnComponent{NumType}, psf_K, ea.S)
-  gal_mcs = Array(GalaxyCacheComponent{NumType}, psf_K, 8, 2, ea.S)
-
-  # TODO: do not keep any derviative information if the sources are not in
-  # active_sources.
-  for s in 1:ea.S
-      psf = ea.patches[s, b].psf
-      vs = ea.vp[s]
-
-      world_loc = vs[[ids.u[1], ids.u[2]]]
-      m_pos = WCSUtils.world_to_pix(ea.patches[s, b].wcs_jacobian,
-                                    ea.patches[s, b].center,
-                                    ea.patches[s, b].pixel_center, world_loc)
-
-      # Convolve the star locations with the PSF.
-      for k in 1:psf_K
-          pc = psf[k]
-          mean_s = [pc.xiBar[1] + m_pos[1], pc.xiBar[2] + m_pos[2]]
-          star_mcs[k, s] =
-            BvnComponent{NumType}(
-              mean_s, pc.tauBar, pc.alphaBar, calculate_siginv_deriv=false)
-      end
-
-      # Convolve the galaxy representations with the PSF.
-      for i = 1:2 # i indexes dev vs exp galaxy types.
-          e_dev_dir = (i == 1) ? 1. : -1.
-          e_dev_i = (i == 1) ? vs[ids.e_dev] : 1. - vs[ids.e_dev]
-
-          # Galaxies of type 1 have 8 components, and type 2 have 6 components.
-          for j in 1:[8,6][i]
-              for k = 1:psf_K
-                  gal_mcs[k, j, i, s] = GalaxyCacheComponent(
-                      e_dev_dir, e_dev_i, galaxy_prototypes[i][j], psf[k],
-                      m_pos, vs[ids.e_axis], vs[ids.e_angle], vs[ids.e_scale],
-                      calculate_derivs && (s in ea.active_sources),
-                      calculate_hessian)
-              end
-          end
-      end
-  end
-
-  star_mcs, gal_mcs
-end
-
-
-"""
-Store pre-allocated memory in this data structures, which contains
-intermediate values used in the ELBO calculation.
-"""
-type HessianSubmatrices{NumType <: Number}
-    u_u::Matrix{NumType}
-    shape_shape::Matrix{NumType}
-end
-
-
-"""
-Pre-allocated memory for efficiently accumulating certain sub-matrices
-of the E_G_s and E_G2_s Hessian.
-
-Args:
-    NumType: The numeric type of the hessian.
-    i: The type of celestial source, from 1:Ia
-"""
-function HessianSubmatrices(NumType::DataType, i::Int)
-    @assert 1 <= i <= Ia
-    shape_p = length(shape_standard_alignment[i])
-
-    u_u = zeros(NumType, 2, 2)
-    shape_shape = zeros(NumType, shape_p, shape_p)
-    HessianSubmatrices{NumType}(u_u, shape_shape)
-end
-
-
-"""
-Add the contributions of a star's bivariate normal term to the ELBO,
-by updating elbo_vars.fs0m_vec[s] in place.
-
-Args:
-    - elbo_vars: Elbo intermediate values.
-    - s: The index of the current source in 1:S
-    - bmc: The component to be added
-    - x: An offset for the component in pixel coordinates (e.g. a pixel location)
-    - wcs_jacobian: The jacobian of the function pixel = F(world) at this location.
-    - calculate_derivs: Whether to calculate derivatives.
-
-Returns:
-    Updates elbo_vars.fs0m_vec[s] in place.
-"""
-function accum_star_pos!{NumType <: Number}(
-                    elbo_vars::ElboIntermediateVariables{NumType},
-                    s::Int,
-                    bmc::BvnComponent{NumType},
-                    x::Vector{Float64},
-                    wcs_jacobian::Array{Float64, 2},
-                    calculate_derivs::Bool)
-    eval_bvn_pdf!(elbo_vars.bvn_derivs, bmc, x)
-
-    # TODO: Also make a version that doesn't calculate any derivatives
-    # if the object isn't in active_sources.
-    get_bvn_derivs!(elbo_vars.bvn_derivs, bmc, true, false)
-
-    fs0m = elbo_vars.fs0m_vec[s]
-    fs0m.v[1] += elbo_vars.bvn_derivs.f_pre[1]
-
-    if elbo_vars.calculate_derivs && calculate_derivs
-        transform_bvn_ux_derivs!(
-            elbo_vars.bvn_derivs, wcs_jacobian, elbo_vars.calculate_hessian)
-        bvn_u_d = elbo_vars.bvn_derivs.bvn_u_d
-        bvn_uu_h = elbo_vars.bvn_derivs.bvn_uu_h
-
-        # Accumulate the derivatives.
-        for u_id in 1:2
-            fs0m.d[star_ids.u[u_id]] += elbo_vars.bvn_derivs.f_pre[1] * bvn_u_d[u_id]
-        end
-
-        if elbo_vars.calculate_hessian
-            # Hessian terms involving only the location parameters.
-            # TODO: redundant term
-            for u_id1 in 1:2, u_id2 in 1:2
-                fs0m.h[star_ids.u[u_id1], star_ids.u[u_id2]] +=
-                    elbo_vars.bvn_derivs.f_pre[1] * (bvn_uu_h[u_id1, u_id2] +
-                    bvn_u_d[u_id1] * bvn_u_d[u_id2])
-            end
-        end
-    end
-end
-
-
-"""
-Add the contributions of a galaxy component term to the ELBO by
-updating fs1m in place.
-
-Args:
-    - elbo_vars: Elbo intermediate variables
-    - s: The index of the current source in 1:S
-    - gcc: The galaxy component to be added
-    - x: An offset for the component in pixel coordinates (e.g. a pixel location)
-    - wcs_jacobian: The jacobian of the function pixel = F(world) at this location.
-    - calculate_derivs: Whether to calculate derivatives.
-
-Returns:
-    Updates elbo_vars.fs1m_vec[s] in place.
-"""
-function accum_galaxy_pos!{NumType <: Number}(
-                    elbo_vars::ElboIntermediateVariables{NumType},
-                    s::Int,
-                    gcc::GalaxyCacheComponent{NumType},
-                    x::Vector{Float64},
-                    wcs_jacobian::Array{Float64, 2},
-                    calculate_derivs::Bool)
-    eval_bvn_pdf!(elbo_vars.bvn_derivs, gcc.bmc, x)
-    f = elbo_vars.bvn_derivs.f_pre[1] * gcc.e_dev_i
-    fs1m = elbo_vars.fs1m_vec[s]
-    fs1m.v[1] += f
-
-    if elbo_vars.calculate_derivs && calculate_derivs
-
-        get_bvn_derivs!(elbo_vars.bvn_derivs, gcc.bmc,
-            elbo_vars.calculate_hessian, elbo_vars.calculate_hessian)
-        transform_bvn_derivs!(
-            elbo_vars.bvn_derivs, gcc.sig_sf, wcs_jacobian, elbo_vars.calculate_hessian)
-
-        bvn_u_d = elbo_vars.bvn_derivs.bvn_u_d
-        bvn_uu_h = elbo_vars.bvn_derivs.bvn_uu_h
-        bvn_s_d = elbo_vars.bvn_derivs.bvn_s_d
-        bvn_ss_h = elbo_vars.bvn_derivs.bvn_ss_h
-        bvn_us_h = elbo_vars.bvn_derivs.bvn_us_h
-
-        # Accumulate the derivatives.
-        for u_id in 1:2
-            fs1m.d[gal_ids.u[u_id]] += f * bvn_u_d[u_id]
-        end
-
-        for gal_id in 1:length(gal_shape_ids)
-            fs1m.d[gal_shape_alignment[gal_id]] += f * bvn_s_d[gal_id]
-        end
-
-        # The e_dev derivative. e_dev just scales the entire component.
-        # The direction is positive or negative depending on whether this
-        # is an exp or dev component.
-        fs1m.d[gal_ids.e_dev] += gcc.e_dev_dir * elbo_vars.bvn_derivs.f_pre[1]
-
-        if elbo_vars.calculate_hessian
-            # The Hessians:
-
-            # Hessian terms involving only the shape parameters.
-            for shape_id1 in 1:length(gal_shape_ids)
-                for shape_id2 in 1:length(gal_shape_ids)
-                    s1 = gal_shape_alignment[shape_id1]
-                    s2 = gal_shape_alignment[shape_id2]
-                    fs1m.h[s1, s2] +=
-                        f * (bvn_ss_h[shape_id1, shape_id2] +
-                                 bvn_s_d[shape_id1] * bvn_s_d[shape_id2])
-                end
-            end
-
-            # Hessian terms involving only the location parameters.
-            for u_id1 in 1:2, u_id2 in 1:2
-                u1 = gal_ids.u[u_id1]
-                u2 = gal_ids.u[u_id2]
-                fs1m.h[u1, u2] +=
-                    f * (bvn_uu_h[u_id1, u_id2] + bvn_u_d[u_id1] * bvn_u_d[u_id2])
-            end
-
-            # Hessian terms involving both the shape and location parameters.
-            for u_id in 1:2, shape_id in 1:length(gal_shape_ids)
-                ui = gal_ids.u[u_id]
-                si = gal_shape_alignment[shape_id]
-                fs1m.h[ui, si] +=
-                    f * (bvn_us_h[u_id, shape_id] + bvn_u_d[u_id] * bvn_s_d[shape_id])
-                fs1m.h[si, ui] = fs1m.h[ui, si]
-            end
-
-            # Do the e_dev hessian terms.
-            devi = gal_ids.e_dev
-            for u_id in 1:2
-                ui = gal_ids.u[u_id]
-                fs1m.h[ui, devi] +=
-                    elbo_vars.bvn_derivs.f_pre[1] * gcc.e_dev_dir * bvn_u_d[u_id]
-                fs1m.h[devi, ui] = fs1m.h[ui, devi]
-            end
-            for shape_id in 1:length(gal_shape_ids)
-                si = gal_shape_alignment[shape_id]
-                fs1m.h[si, devi] +=
-                    elbo_vars.bvn_derivs.f_pre[1] * gcc.e_dev_dir * bvn_s_d[shape_id]
-                fs1m.h[devi, si] = fs1m.h[si, devi]
-            end
-        end # if calculate hessian
-    end # if calculate_derivs
-end
